ai_business_decision_copilot/requirements.txt

### `requirements.txt`

```txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
openai>=1.0.0
python-dotenv>=1.0.0
prophet>=1.1.5